pyproject.toml
[project]
name = "web-python-labs"
version = "0.1.0"
description = ""
authors = [
    {name = "futurezayka",email = "kurachvlad911@gmail.com"}
]

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.115.0"}
loguru = "^0.7.3"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
pydantic = {version = "^2.10.3", extras = ["email"]}
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic-extra-types = "^2.10.1"
python-jose = "^3.4.0"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

# Assume Python 3.11
target-version = "py312"

# Allow fix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
